+++ CHANGES.rst
@@ -37,6 +37,8 @@ Unreleased
     ``python -m name``. :issue:`1603`
 -   Include required parent arguments in help synopsis of subcommands.
     :issue:`1475`
+-   Help for boolean flags with ``show_default=True`` shows the flag
+    name instead of ``True`` or ``False``. :issue:`1538`
 
 
 Version 7.1.2